ServoControl.ino
This is a work in progress. Right now, the PID control factors are likely too large or the way I try to correct for error is flawed which leads to the motors overcorrecting and making the motors move far from the subject. I will upload the correct version when I sort these issues out.