Add files via upload
This includes mounting brackets for the X and Y axis servo motors as well as a modified Raspberry Pi Case that has a passthrough for the PiCamera 2 ribbon cable. The screw widths for the mounting brackets and baseplate is the size of the standard screw from the Muizei MS18 servo motor.